feat: /RegistrationForm add request to user what role it choose
