+++ src/components/RegistrationForm/components/SponsorForm/index.js
@@ -1,0 +1,1 @@
+export { SponsorForm } from './SponsorForm';
