+++ src/components/RegistrationForm/components/TalentForm/index.js
@@ -1,0 +1,1 @@
+export { TalentForm } from './TalentForm';
