+++ src/components/RegistrationForm/components/TalentForm/validation.js
